requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
Django==4.0.3
django-cors-headers==3.11.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
executing==0.8.3
frozenlist==1.3.0
h5py==3.6.0
halo==0.0.31
idna==3.3
ipython==8.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
log-symbols==0.0.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
numpy==1.22.3
openapi-codec==1.3.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python-weather==0.3.7
pytz==2022.1
requests==2.27.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
spinners==0.0.24
sqlparse==0.4.2
stack-data==0.2.0
termcolor==1.1.0
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.0.0
xmltodict==0.12.0
yarl==1.7.2
